Increase code readability of src files. (#63)
